+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Feature Summary
+- [check editor's markers]
 - [demos to using angular-monaco-editor component]
 - [Init monaco editor using angular]
+
+# 1.7.0
+  Extract model as a service from the angular-monaco-editor component.
 
 # 1.6.0
   Add model markers actived entrance.